[CImagePlugin] Update to new typeinfo system.
